Move api key parsing to conf module
