+++ src/main.rs
@@ -1,28 +1,13 @@
-
-use std::env;
-use std::process;
 use geojson::GeoJson;
 use geojson::Value::Point;
 
 use crate::route::Route;
 use crate::step::Step;
-use crate::conf::Args;
 
 mod route;
 mod step;
 mod conf;
 
-fn get_api_key() -> String {
-    dotenv::dotenv().ok();
-    let key = "ROUTESERVICE_API_KEY";
-    match env::var(key) {
-        Ok(val) => val,
-        Err(e) => {
-            println!("You should define {}. Following error happened: {}", key, e);
-            process::exit(1);
-        }
-    }
-}
 
 fn parse_geolocation(geo_result_string: &str)  -> Option<(f64, f64)> {
     if let GeoJson::FeatureCollection(parsed_geo_result) = geo_result_string.parse::<GeoJson>().unwrap() {
@@ -56,8 +41,9 @@ async fn get_directions(profile: &str, start: (f64, f64), end: (f64, f64), api_k
 
 #[tokio::main]
 async fn main() {
-    let (start, end, profile) = Args::get_args();
-    let api_key = get_api_key();
+    let (start, end, profile) = conf::Args::get_args();
+    let api_key = conf::get_api_key();
+
     let start_georesult = get_geolocation(&start, &api_key).await.unwrap();
     let end_georesult = get_geolocation(&end, &api_key).await.unwrap();
